notadjango/requirements.txt
django
Pillow
mysqlclient
django-autoslug